Изменена структура БД. +Много мелких правок и доработок.
Добавлен функционал смены пользователей в клиенте.
Добавлено шифрование паролей.
Добавлено разграничение доступа пользователей.
! Изменена структура БД +добавлены триггеры.
Частично убрана статистика.
Переписан отчет по возвращенным билетам в админке.
Добавлены предупредительные сообщения в админке при критических
изменениях.

+++ Admin/MainManager/actionsWidget/cactionswidget.cpp
@@ -103,16 +103,25 @@ void CActionsWidget::on_tbnDel_clicked()
 	QTreeWidgetItem *selectedItem = ui->twActions->currentItem();
 	if(selectedItem)
 	{
-		if(QMessageBox::Yes == QMessageBox::question(this, tr("Запрос подтверждения"), tr("Вы действительно хотите удалить выбранное мероприятие?"), QMessageBox::Yes, QMessageBox::No))
+		QSqlQuery query(QSqlDatabase::database(mConnectionName));
+		query.prepare("SELECT COUNT(Tickets.id), COUNT(Reservations.id) FROM Tickets, Reservations WHERE Tickets.id_action = :id AND Reservations.id_action = :id2;");
+		query.bindValue(":id", selectedItem->text(ID));
+		query.bindValue(":id2", selectedItem->text(ID));
+		if(query.exec() && query.first())
 		{
-			QSqlQuery query(QSqlDatabase::database(mConnectionName));
-			query.prepare("DELETE FROM Actions WHERE id = :id");
-			query.bindValue(":id", selectedItem->text(ID));
-			query.exec();
-
-			////TODO: Сделать каскадное удаление.
-
-			updateData();
+			if(query.value(0).toInt() + query.value(1).toInt() > 0)
+			{
+				QMessageBox::warning(this, tr("Внимание"), tr("Невозможно удалить данное мероприятие так как уже продано %1 и забронировано %2 билетов.\nУдаление мероприятия невозможно. Но Вы можете изменить его состояние на '%3' - это равносильно удалению.").arg(query.value(0).toString()).arg(query.value(1).toString()).arg(Global::actionStateToText(Global::ActionComplited)));
+				return;
+			}
+			else
+				if(QMessageBox::Yes == QMessageBox::question(this, tr("Запрос подтверждения"), tr("Вы действительно хотите удалить выбранное мероприятие?"), QMessageBox::Yes, QMessageBox::No))
+				{
+					query.prepare("DELETE FROM Actions WHERE id = :id");
+					query.bindValue(":id", selectedItem->text(ID));
+					query.exec();
+					updateData();
+				}
 		}
 	}
 }
